% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_genes_for_go_term.R
\name{extract_genes_for_go_term}
\alias{extract_genes_for_go_term}
\title{Extract genes of interest associated with a specific GO term}
\usage{
extract_genes_for_go_term(GOdata, go_term_of_interest, gene_list)
}
\arguments{
\item{GOdata}{A `topGOdata` object created during the GO enrichment analysis.}

\item{go_term_of_interest}{A character string specifying the GO term (GO.ID) of interest.}

\item{gene_list}{A character vector of gene names that represents your genes of interest.}
}
\value{
A character vector containing the genes from `gene_list` associated with the specified GO term.
}
\description{
This function extracts the genes from your `gene_list` that are associated with a given GO term (GO.ID)
from the `topGOdata` object generated during a GO enrichment analysis.
}
\examples{
# Assuming `GOdata` is your topGOdata object and `gene_list` is your list of genes:
# genes_of_interest <- extract_genes_for_go_term(GOdata, "GO:0030941", gene_list)

}
