% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_in_features.R
\name{features_in_features}
\alias{features_in_features}
\title{Overlap of Genomic Features}
\usage{
features_in_features(features, features2, mode, value = NULL)
}
\arguments{
\item{features}{A `data.table` containing the first set of genomic features.
It must include the columns "CHROM", "START", "STOP", and "ID".}

\item{features2}{A second `data.table` containing the second set of genomic features.
It must include the columns "CHROM", "START", and "STOP".}

\item{mode}{The mode of calculation: "counts" for number of overlaps,
"any" for any overlaps (TRUE/FALSE), "mean" for mean of a specified column in the features2 data.table,
"sum" for the sum of a specified column, "length" for the total length of the overlap,
"sumxlength" for the sum of a specified column multiplied by the length of the overlap.}

\item{value}{The name of a column in the features2 data.table to be used when mode = "mean", "sum", or "sumxlength".}
}
\value{
A `data.table` with the results of the calculation for each feature in the first set.
        The result column is named after the 'value' parameter if 'mode' is 'mean', 'sum', or 'sumxlength', or 'overlaps' if 'mode' is 'counts' or 'any'.
}
\description{
This function calculates the overlap between two sets of genomic features.
It allows the user to compute counts of overlaps, determine whether any overlap exists,
and calculate the mean, sum, length, or sum multiplied by length of a specified value in the second genomic feature set.
}
\examples{
# Ensure features and features2 data.tables are properly structured before using this function.
# Example:
# features <- data.table(CHROM = c("chr1", "chr1", "chr2"), START = c(1, 100, 200), STOP = c(50, 150, 250), ID = c("feat1", "feat2", "feat3"))
# features2 <- data.table(CHROM = c("chr1", "chr1", "chr2"), START = c(25, 125, 225), STOP = c(75, 175, 275))
# features_in_features(features, features2, mode = "counts")

}
\seealso{
\code{\link[data.table]{data.table}}, \code{\link[data.table]{foverlaps}}
}
\keyword{genomics}
