% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_Nmer_frequencies.R
\name{genome_Nmer_frequencies}
\alias{genome_Nmer_frequencies}
\title{Calculate N-mer Frequencies in Genomic Sequences}
\usage{
genome_Nmer_frequencies(genome, Nmer, stop = NULL)
}
\arguments{
\item{genome}{A FASTA object as read by `seqinr::read.fasta()`. This should be a list where
each element represents a chromosome or sequence, with names corresponding to chromosome
or sequence identifiers.}

\item{Nmer}{An integer specifying the size of the N-mer to analyze.}

\item{stop}{An optional integer. If provided, the analysis is limited to the first `stop`
positions of each sequence. By default (`NULL`), the entire length of each sequence is
considered.}
}
\value{
A data.table with two columns: `context`, representing the unique N-mers identified across
        all sequences, and `N`, representing the sum of occurrences of each N-mer. This table
        provides a comprehensive overview of N-mer frequencies within the specified genomic regions.
}
\description{
This function calculates the frequencies of N-mers (specified by the user) in genomic sequences
provided as a FASTA object. It is designed to efficiently process large genomic sequences by
optionally limiting the analysis to the first `stop` positions of each chromosome. The results
are summarized in a data.table showing the counts of each N-mer across all processed sequences.
}
\examples{
# Assuming `fasta_object` is read using seqinr::read.fasta()
Nmer_counts <- calculate_Nmer_frequencies(fasta_object, Nmer = 3, stop = 10000)

}
