% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homopolymer_var_annotate.R
\name{homopolymer_var_annotate}
\alias{homopolymer_var_annotate}
\title{Annotate Variants Near Homopolymers}
\usage{
homopolymer_var_annotate(sites, homopolymers, size, dist)
}
\arguments{
\item{sites}{A data.table with CHROM, POS, REF, ALT columns representing the genomic sites.}

\item{homopolymers}{A data.table returned from the `make_homopolymer()` function, representing homopolymer regions by CHROM.}

\item{size}{The minimum length of homopolymers to consider.}

\item{dist}{The maximum distance a variant can be from a homopolymer to be considered a neighbor.}
}
\value{
A data.table with a new column, 'homopolymer_neighbor', which indicates whether each site is a neighbor to a homopolymer.
}
\description{
This function identifies and annotates variants that are in or near homopolymer runs in the genome.
Variants are annotated based on whether they are located within a certain distance (specified by 'dist')
from a homopolymer run of a certain size (specified by 'size').
}
\examples{
# Assume 'sites' and 'homopolymers' are pre-defined
# sites <- data.table( ... )
# homopolymers <- make_homopolymer( ... )
# Annotate sites
sites <- homopolymer_var_annotate(sites, homopolymers, 5, 10)

}
\seealso{
\code{\link[data.table]{foverlaps}}, \code{\link[data.table]{rbindlist}}, \code{\link[data.table]{setkey}}
}
\keyword{genomics}
