% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_codon_table.R
\name{make_codon_table}
\alias{make_codon_table}
\title{Create a Standard Genetic Codon Table}
\usage{
make_codon_table()
}
\value{
A data.table containing the following columns:
\itemize{
  \item \strong{AA}: Amino acid corresponding to the codon.
  \item \strong{codon}: The codon sequence.
  \item \strong{A}: Count of adenine in the codon.
  \item \strong{T}: Count of thymine in the codon.
  \item \strong{C}: Count of cytosine in the codon.
  \item \strong{G}: Count of guanine in the codon.
}
}
\description{
This function generates a data table representing the standard genetic code.
The table maps codons to their corresponding amino acids and also provides
the count of each nucleotide (A, T, C, G) in the codon.
}
\examples{
codon_tbl <- make_codon_table()
head(codon_tbl)

}
