% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_homopolymer.R
\name{make_homopolymer}
\alias{make_homopolymer}
\title{Identify Homopolymer Runs}
\usage{
make_homopolymer(fasta, size)
}
\arguments{
\item{fasta}{A named list representing the genome sequence.}

\item{size}{The minimum length of homopolymers to identify. This parameter
is passed to the `homopolymer_read` function.}
}
\value{
A list of identified homopolymers for each sequence in the genome.
}
\description{
This function is a wrapper for the `homopolymer_read` function. It is used
to identify runs of homopolymers in a genome sequence.
}
\examples{
# Define a list representing a genome sequence
genome_seq <- list("seq1" = c("A", "A", "A", "T", "G", "G", "G", "G"),
                   "seq2" = c("C", "C", "C", "A", "A", "T", "T", "T"))
# Call the function to identify homopolymers
homopolymers <- make_homopolymer(genome_seq, 2)

}
\seealso{
\code{\link[homopolymer_read]{homopolymer_read}}
}
\keyword{genomics}
