% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nmer_protein_aa.R
\name{Nmer_protein_aa}
\alias{Nmer_protein_aa}
\title{Nmer_protein_aa}
\usage{
Nmer_protein_aa(proteins, Nmer, mode)
}
\arguments{
\item{proteins}{A list of protein sequences read by `read.fasta` from the `seqinr` package.}

\item{Nmer}{An integer specifying the length of the amino acid motifs.}

\item{mode}{A character string, either "prop" for proportions or "counts" for raw counts.}
}
\value{
A data.table where each row represents a protein and columns reflect N-mer frequencies.
}
\description{
This function calculates the frequency or proportion of N-mer amino acid motifs in a set of proteins.
}
\examples{
proteins <- read.fasta("~/Documents/TAIR10_pep_20101214_updated")
Nmer_protein_aa(proteins, Nmer = 2, mode = "counts")
}
