% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nmerfrequency.R
\name{Nmerfrequency}
\alias{Nmerfrequency}
\title{N-mer Frequency}
\usage{
Nmerfrequency(features, fasta, Nmer, mode = "prop")
}
\arguments{
\item{features}{A `data.table` containing genomic features. It must include the columns "CHROM", "START", "STOP", and "ID".}

\item{fasta}{A named list representing the genome sequence. Names of the list should correspond to chromosome names, and each item in the list should be a character vector representing the sequence of the corresponding chromosome.}

\item{Nmer}{numeric value (2 would give dimers, 3 would give trimers, etc)}

\item{mode}{a string that can be either "prop" or "counts" to return either Nmer frequencies as total numbers or proportions.}
}
\value{
A `data.table` containing the frequency of each N-mer for each genomic feature.
}
\description{
Calculate the frequency of N-mers in given genomic features.
}
