% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonsyn_syn_exp2.R
\name{nonsyn_syn_exp2}
\alias{nonsyn_syn_exp2}
\title{Nonsynonymous to Synonymous Mutation Expectation}
\usage{
nonsyn_syn_exp2(muts, genome, CDS, reps, n)
}
\arguments{
\item{muts}{A data.table with REF and ALT columns indicating the reference and alternate nucleotides.}

\item{CDS}{A coding sequence fasta object read by seqinr.}

\item{reps}{Number of repetitions for the simulation.}

\item{n}{Number of samples per repetition.}
}
\value{
A list containing the results for each repetition and the overall ratio of nonsynonymous to synonymous mutations.
}
\description{
This function calculates the expected ratio of nonsynonymous to synonymous mutations
based on a given mutation dataset and a coding sequence.
}
\examples{
muts <- fread("~/Dropbox/Research/rice mutation paper/data/Strelka2_mutations.csv")[trt=="MSH6"]
CDS <- read.fasta("~/Documents/TAIR10_cds_20110103_representative_gene_model_updated.txt")
result <- nonsyn_syn_exp(muts, CDS, 10, 1000)


}
