% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_GFF_INFO.R
\name{parse_GFF_INFO}
\alias{parse_GFF_INFO}
\title{Parse INFO Column from GFF Object}
\usage{
parse_GFF_INFO(GFF)
}
\arguments{
\item{GFF}{A data frame or data.table representing the GFF object. It must contain an INFO column.}
}
\value{
The GFF object with the INFO column parsed into new columns for all values.
}
\description{
This wrapper function runs `parse_INFO` on every value of the INFO column of a GFF object.
It takes a GFF object, parses the INFO column, and adds new columns to the object for all the parsed values.
GFF is an object returned by read.GFF() and must contain an INFO column.
}
\examples{
GFF <- read.GFF("path/to/your.gff")
parsed_GFF <- parse_GFF_INFO(GFF)
}
