% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_INFO.R
\name{parse_INFO}
\alias{parse_INFO}
\title{Parse INFO Field from GFF Format}
\usage{
parse_INFO(INFO)
}
\arguments{
\item{INFO}{A character string representing the INFO field from a GFF file. It can be obtained using read.GFF()$INFO.}
}
\value{
A data.table containing the parsed key-value pairs from the INFO field.
}
\description{
This function takes a single INFO field from the GFF (General Feature Format) file
and parses it into a data table. The INFO field consists of a semicolon-separated
string of key-value pairs, where each key and value are separated by an equal sign.
Note: This is a subfunction run by parse_GFF_INFO.
}
\examples{
info_string <- "key1=value1;key2=value2;key3=value3"
parsed_info <- parse_INFO(info_string)
}
