% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_topGO_analysis.R
\name{perform_topGO_analysis}
\alias{perform_topGO_analysis}
\title{Perform GO enrichment analysis using the topGO package}
\usage{
perform_topGO_analysis(gene_go_table, gene_list, ontology = "MF")
}
\arguments{
\item{gene_go_table}{A `data.table` containing two columns:
\itemize{
  \item `GENE`: Gene identifiers.
  \item `GO`: Corresponding Gene Ontology (GO) terms.
}}

\item{gene_list}{A character vector of gene names (in the same format as the `GENE` column of `gene_go_table`)
representing the genes of interest (e.g., differentially expressed genes).}

\item{ontology}{A character string specifying the ontology to use.
Options are "MF" (Molecular Function), "BP" (Biological Process), or "CC" (Cellular Component).
Default is "MF".}
}
\value{
A data.frame containing the top enriched GO terms with both raw and adjusted p-values
  (Benjamini-Hochberg corrected), for both the classic and elim algorithms.
}
\description{
This function performs a Gene Ontology (GO) enrichment analysis using the `topGO` package.
It tests for GO term enrichment in a list of genes compared to a background set of genes.
}
\examples{
# Assuming gene_go_table is a data.table of gene-to-GO mappings and gene_list is a vector of genes of interest:
# result <- perform_topGO_analysis(gene_go_table, gene_list, ontology = "BP")

}
