% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bins.R
\name{plot_bins}
\alias{plot_bins}
\title{plot_bins Function}
\usage{
plot_bins(data, yvar, xvar, bins)
}
\arguments{
\item{data}{A data.table containing the data.}

\item{yvar}{Character name of the column in data to be used as the y-variable.}

\item{xvar}{Character name of the column in data to be used as the x-variable.}

\item{bins}{Number of bins to divide the x-variable into.}
}
\value{
A list containing the binned data and the ggplot object.
}
\description{
This function takes a data.table, two column names, and a number of bins as input.
It then calculates the mean and standard error of the y-variable for each bin of the x-variable.
Finally, it returns a ggplot object showing the binned data.
}
\examples{
\dontrun{
library(data.table)
dt <- data.table(enrich = rnorm(100), g1001 = rnorm(100))
result <- plot_bins(dt, "g1001", "enrich", 10)
print(result[[2]])
}

}
