% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_feature_windows.R
\name{plot_feature_windows}
\alias{plot_feature_windows}
\title{Plot Variable in Relation to Features}
\usage{
plot_feature_windows(feature_windows, variable, mode)
}
\arguments{
\item{feature_windows}{A `data.table` object created by the `feature_windows` function. It should include the columns "REGION", "RELATIVEPOS", and "LENGTH".}

\item{variable}{A string specifying the name of the column in the `feature_windows` data.table to be plotted.}

\item{mode}{A string specifying the mode of calculation: "mean" for mean of the variable, "percent" for the sum of the variable values divided by the total length.}
}
\value{
A `ggplot` object representing the plot of the variable in relation to features.
}
\description{
This function generates a plot of a specified variable in relation to features in a set of genomic feature windows.
It allows for calculation and plotting of the mean or percentage of a variable.
}
\examples{
# Ensure 'feature_windows' data.table is properly structured before using this function.
# Example:
# feature_windows <- feature_windows(features = features, chipfile = chipfile, inputfile = inputfile, mode = "mean", value="DEPTH")
# plot_feature_windows(feature_windows = feature_windows, variable = "DEPTH", mode = "mean")

}
\seealso{
\code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{geom_line}}, \code{\link[ggplot2]{geom_vline}}, \code{\link[ggplot2]{theme_classic}}, \code{\link[ggplot2]{scale_x_continuous}}
}
\keyword{genomics,}
\keyword{plotting}
