% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.R
\name{qqplot}
\alias{qqplot}
\title{QQ Plot for P-values}
\usage{
qqplot(p_values)
}
\arguments{
\item{p_values}{A numeric vector of p-values to be plotted.}
}
\value{
A ggplot object representing the QQ plot.
}
\description{
Generates a QQ plot to visualize the distribution of observed versus expected -log10 transformed p-values.
}
\details{
This function calculates the expected and observed -log10(p-values) and
plots them on a QQ plot. The diagonal line represents the null hypothesis of
no association (y = x). Points deviating from the line suggest potential associations.
}
\examples{
# Example usage:
p_values <- runif(100, min = 1e-8, max = 1)
qqplot(p_values)

}
