% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.bedGraph.R
\name{read.bedGraph}
\alias{read.bedGraph}
\title{Read bedGraph files}
\usage{
read.bedGraph(file)
}
\arguments{
\item{file}{A character string specifying the path to the bedGraph file.}
}
\value{
A `data.table` with the bedGraph data, where columns are named "CHROM", "START", "STOP", and "DEPTH".
}
\description{
This function reads in a bedGraph file, assigns standard column names
and checks for the correct format.
}
\examples{
# Assuming 'example.bedGraph' is a correctly formatted bedGraph file in your working directory:
# data <- read.bedGraph('example.bedGraph')

}
\keyword{fileIO}
\keyword{genomics}
