% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.GFF.R
\name{read.GFF}
\alias{read.GFF}
\title{Read a GFF Formatted File}
\usage{
read.GFF(file_path)
}
\arguments{
\item{file_path}{The path to the GFF formatted file to read in.}
}
\value{
A data.table with named columns from the GFF file.
}
\description{
This function reads a GFF formatted file and assigns the column names.
It also checks that the object has the expected number of columns and
that START and STOP are numeric.
}
\examples{
# Ensure the GFF file is in the correct directory before using this function.
}
