% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.VCF.R
\name{read.VCF}
\alias{read.VCF}
\title{Read and parse a VCF file}
\usage{
read.VCF(file, caller)
}
\arguments{
\item{file}{Path to the VCF file to be read.}

\item{caller}{Name of the caller used. Options are "strelka2", "haplotypecaller", "pbsv","other".}
}
\value{
A data.table containing the fixed fields and genotype information from the VCF file.
If the caller is "strelka2", an additional EVS column is added to the output.
}
\description{
This function uses the read.vcfR function from the vcfR package to read and parse a VCF file.
It allows for input from different callers including Strelka2, HaplotypeCaller, and PBSV.
}
\examples{
# Ensure vcfR package is installed before using this function.
# read.VCF("path_to_your_file.vcf", "strelka2")

}
