% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_blast.R
\name{read_blast}
\alias{read_blast}
\title{Read BLAST Output}
\usage{
read_blast(file)
}
\arguments{
\item{file}{A string. The file path corresponding to the output of a BLAST search in fmt 6.}
}
\value{
A data.table object containing BLAST results with appropriate column names.
}
\description{
This function reads BLAST output in format 6 and returns the results in a data.table format with appropriate column names.
}
\examples{
# Assuming 'sample_blast_output.txt' is a BLAST output in fmt 6
blast_results <- read_blast('sample_blast_output.txt')

}
