% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_sci_format.R
\name{real_sci_format}
\alias{real_sci_format}
\title{Format numbers with real scientific notation}
\usage{
real_sci_format(x, signif_N = NULL)
}
\arguments{
\item{x}{A numeric vector of values to be formatted.}

\item{signif_N}{A numeric value for rounding small values to signif_N digits with signif, default is NULL}
}
\value{
An expression vector with numbers formatted in scientific notation suitable for plot rendering.
}
\description{
This function formats numeric values into a scientific notation with true exponents.
For example, instead of '1e-4', the output would render as \(1 \times 10^{-4}\) on a plot.
}
\examples{
real_sci_format(c(1000, 0.0001, 1e-7))
}
