% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revcomp.R
\name{revcomp}
\alias{revcomp}
\title{Reverse Complement of a DNA Sequence}
\usage{
revcomp(motif)
}
\arguments{
\item{motif}{A character string representing the DNA sequence for which the reverse complement is desired.}
}
\value{
A character string representing the reverse complement of the input DNA sequence.
}
\description{
This function returns the reverse complement of a given DNA sequence (motif).
It utilizes the `comp` function from the `seqinr` package to generate the complement.
}
\examples{
revcomp("ATGC")
# Returns "GCAT"

}
