% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites_in_features.R
\name{sites_in_features}
\alias{sites_in_features}
\title{Identify Overlapping Sites in Features}
\usage{
sites_in_features(features, sites, mode, value = NULL)
}
\arguments{
\item{features}{A data.table containing genomic features. Must include "CHROM", "START", "STOP", and "ID" columns.}

\item{sites}{A data.table of sites (mutations). Must include "CHROM", "POS" columns.}

\item{mode}{A character string indicating the mode of overlap calculation. Must be one of "counts", "any", or "mean".}

\item{value}{A character string indicating the column name in 'sites' data.table to calculate the mean. Only required when mode = "mean".}
}
\value{
A data.table summarizing the overlaps for each feature (identified by "ID"), according to the specified mode.
}
\description{
This function identifies overlapping sites from a provided data.table within given genomic features.
The function expects "CHROM", "START", "STOP", and "ID" columns in the features data.table, and
"CHROM", "POS" columns in the sites data.table. The 'mode' parameter controls the method
of summarizing overlaps: 'counts' gives the total count, 'any' gives a Boolean value
indicating the presence of any overlap, and 'mean' calculates the mean of a specified 'value' column.
}
\examples{
# Ensure that features and sites data.tables are properly structured before using this function.
# Example:
# features <- data.table(CHROM = c("chr1", "chr1", "chr2"), START = c(1, 100, 200), STOP = c(50, 150, 250), ID = c("feat1", "feat2", "feat3"))
# sites <- data.table(CHROM = c("chr1", "chr1", "chr2"), POS = c(25, 125, 225))
# sites_in_features(features, sites, mode = "counts")

}
\seealso{
\code{\link[data.table]{data.table}}, \code{\link[data.table]{foverlaps}}
}
\keyword{genomics}
