% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strelka2_filter.R
\name{strelka2_filter}
\alias{strelka2_filter}
\title{strelka2_filter}
\usage{
strelka2_filter(directory, output)
}
\arguments{
\item{directory}{A string specifying the path to the directory containing the VCF files.}

\item{output}{A string specifying the path to the output directory.}
}
\value{
The function writes two CSV files in the output directory: "strelka2_all_calls.csv" and "strelka2_filtered_calls.csv".
}
\description{
This function reads in results from strelka2 VCF files and applies various filters to identify de novo mutations.
}
