% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricontexts.R
\name{tricontexts}
\alias{tricontexts}
\title{Trinucleotide context}
\usage{
tricontexts(sites, fasta)
}
\arguments{
\item{sites}{A data.table of mutations. It should contain "CHROM", "POS",
"REF", and "ALT" columns, representing chromosome, position, reference allele,
and alternate allele, respectively.}

\item{fasta}{A named list representing the genome sequence. Names of the list
should correspond to chromosome names, and each item in the list should be a
character vector representing the sequence of the corresponding chromosome.}
}
\value{
A vector of contexts for each mutation.
}
\description{
This function obtains the trinucleotide context for a given set of mutations.
The context is determined by examining the base pairs immediately flanking
the mutation site.
}
