% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimer_mutation_rate.R
\name{trimer_mutation_rate}
\alias{trimer_mutation_rate}
\title{Trimer Mutation Rate}
\usage{
trimer_mutation_rate(tricontexts, trimer_counts)
}
\arguments{
\item{tricontexts}{A vector generated from the `tricontexts()` function.}

\item{trimer_counts}{A `data.table` generated from the `trimerfrequencycomp(Nmerfrequency(Nmer=3, mode="counts"))` function.}
}
\value{
A `data.table` containing the mutation rate for each trimer.
}
\description{
Calculate the mutation rate for each trimer.
}
\examples{
trimer_mutation_rate(tricontexts, trimer_counts)
}
