% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimerfrequencycomp.R
\name{trimerfrequencycomp}
\alias{trimerfrequencycomp}
\title{Trimer Frequency Comparison}
\usage{
trimerfrequencycomp(Nmerfrequencytable)
}
\arguments{
\item{Nmerfrequencytable}{A data.table generated by Nmerfrequency with Nmer=3}
}
\value{
A data.table with summed trimer frequencies for complements
}
\description{
This function computes trimer frequency comparison for given Nmer frequency table.
}
\examples{
# Assuming `Nmerfrequencytable` is already defined
trimerfrequencycomp(Nmerfrequencytable)

}
