% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_hunter.R
\name{motif_hunter}
\alias{motif_hunter}
\title{Motif Hunter in Fasta Object}
\usage{
motif_hunter(fasta, motif)
}
\arguments{
\item{fasta}{A fasta object as read by read.fasta from the seqinr package in R. Must be a list.}

\item{motif}{A DNA sequence. Can be upper or lower case but needs to match the case of the fasta object.}
}
\value{
A data.table with locations of the motif within the fasta object.
}
\description{
This function hunts for a specified DNA sequence motif in a given fasta object.
The fasta object must match the case (upper vs lower) of the motif.
}
\examples{
# Ensure fasta object and motif are properly structured before using this function.
# motif_hunter(fasta, motif)

}
